requirements/local.txt
# Local development dependencies go here
-r base.txt

coverage==4.0.3
django-debug-toolbar==1.4
Sphinx==1.2.2
